Fixed bug in task 7

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,11 +8,11 @@
  */
 int print_last_digit(int n)
 {
-	int r = n;
-
+	long int r = n;
 
 	if (r < 0)
 		r = -1 * r;
+
 	r = r % 10;
 	_putchar('0' + r);
 	return (r);
